+++ CHANGELOG.rst
@@ -77,6 +77,7 @@ Changed
     - Return ``500`` for iRODS query exceptions in ``ProjectIrodsFileListAPIView`` (#2103)
     - Exclude inactive users from email sending and alert creation (#2114)
     - Display disabled path field in iRODS access ticket update form (#2139)
+    - Allow iRODS access ticket creation for data objects in UI (#2138)
 - **Taskflowbackend**
     - Enable no role for old owner in ``perform_owner_transfer()`` (#2051)
     - Rename ``BatchCheckFileTask`` to ``BatchCheckFileExistTask`` (#2064)